Break FieldValue and Timestamp into separate files
