+++ mocks/firestore.js
@@ -9,22 +9,13 @@ const mockAdd = jest.fn();
 const mockSet = jest.fn();
 const mockDelete = jest.fn();
 
-const mockArrayRemoveFieldValue = jest.fn();
-const mockArrayUnionFieldValue = jest.fn();
-const mockDeleteFieldValue = jest.fn();
-const mockIncrementFieldValue = jest.fn();
-const mockServerTimestampFieldValue = jest.fn();
-
 const mockBatchDelete = jest.fn();
 const mockBatchCommit = jest.fn();
 const mockBatchUpdate = jest.fn();
 const mockBatchSet = jest.fn();
 
-const mockTimestampToDate = jest.fn();
-const mockTimestampToMillis = jest.fn();
-const mockTimestampFromDate = jest.fn();
-const mockTimestampFromMillis = jest.fn();
-
+const timestamp = require('./timestamp');
+const fieldValue = require('./fieldValue');
 const query = require('./query');
 const transaction = require('./transaction');
 
@@ -186,133 +177,8 @@ class FakeFirestore {
 
 FakeFirestore.Query = query.Query;
 FakeFirestore.Transaction = transaction.Transaction;
-
-FakeFirestore.FieldValue = class {
-  constructor(type, value) {
-    this.type = type;
-    this.value = value;
-  }
-
-  isEqual(other) {
-    return (
-      other instanceof FakeFirestore.FieldValue &&
-      other.type === this.type &&
-      other.value === this.value
-    );
-  }
-
-  transform(value) {
-    switch (this.type) {
-      case 'arrayUnion':
-        if (Array.isArray(value)) {
-          return value.concat(this.value.filter(v => !value.includes(v)));
-        } else {
-          return this.value;
-        }
-      case 'arrayRemove':
-        if (Array.isArray(value)) {
-          return value.filter(v => !this.value.includes(v));
-        } else {
-          return value;
-        }
-      case 'increment': {
-        const amount = Number(this.value);
-        if (typeof value === 'number') {
-          return value + amount;
-        } else {
-          return amount;
-        }
-      }
-      case 'serverTimestamp': {
-        return FakeFirestore.Timestamp.now();
-      }
-      case 'delete':
-        return undefined;
-    }
-  }
-
-  static arrayUnion(elements = []) {
-    mockArrayUnionFieldValue(...arguments);
-    if (!Array.isArray(elements)) {
-      elements = [elements];
-    }
-    return new FakeFirestore.FieldValue('arrayUnion', elements);
-  }
-
-  static arrayRemove(elements) {
-    mockArrayRemoveFieldValue(...arguments);
-    if (!Array.isArray(elements)) {
-      elements = [elements];
-    }
-    return new FakeFirestore.FieldValue('arrayRemove', elements);
-  }
-
-  static increment(amount = 1) {
-    mockIncrementFieldValue(...arguments);
-    return new FakeFirestore.FieldValue('increment', amount);
-  }
-
-  static serverTimestamp() {
-    mockServerTimestampFieldValue(...arguments);
-    return new FakeFirestore.FieldValue('serverTimestamp');
-  }
-
-  static delete() {
-    mockDeleteFieldValue(...arguments);
-    return new FakeFirestore.FieldValue('delete');
-  }
-};
-
-FakeFirestore.Timestamp = class {
-  constructor(seconds, nanoseconds) {
-    this.seconds = seconds;
-    this.nanoseconds = nanoseconds;
-  }
-
-  isEqual(other) {
-    return (
-      other instanceof FakeFirestore.Timestamp &&
-      other.seconds === this.seconds &&
-      other.nanoseconds === this.nanoseconds
-    );
-  }
-
-  toDate() {
-    const d = new Date(0);
-    d.setSeconds(this.seconds);
-    d.setMilliseconds(this.nanoseconds / 1000000);
-    return mockTimestampToDate(...arguments) || d;
-  }
-
-  toMillis() {
-    const d = new Date(0);
-    d.setSeconds(this.seconds);
-    d.setMilliseconds(this.nanoseconds / 1000000);
-    return mockTimestampToMillis(...arguments) || d.getMilliseconds();
-  }
-
-  valueOf() {
-    return JSON.stringify(this.toMillis());
-  }
-
-  static fromDate(date) {
-    return (
-      mockTimestampFromDate(...arguments) ||
-      new FakeFirestore.Timestamp(date.getSeconds(), date.getMilliseconds() * 1000000)
-    );
-  }
-
-  static fromMillis(millis) {
-    const d = new Date(0);
-    d.setMilliseconds(millis);
-    return mockTimestampFromMillis(...arguments) || FakeFirestore.Timestamp.fromDate(d);
-  }
-
-  static now() {
-    const now = new Date();
-    return FakeFirestore.Timestamp.fromDate(now);
-  }
-};
+FakeFirestore.FieldValue = fieldValue.FieldValue;
+FakeFirestore.Timestamp = timestamp.Timestamp;
 
 module.exports = {
   FakeFirestore,
@@ -326,23 +192,13 @@ module.exports = {
   mockGetAll,
   mockSet,
   mockUpdate,
-  mockArrayRemoveFieldValue,
-  mockArrayUnionFieldValue,
-  mockDeleteFieldValue,
-  mockIncrementFieldValue,
-  mockServerTimestampFieldValue,
   mockBatchDelete,
   mockBatchCommit,
   mockBatchUpdate,
   mockBatchSet,
-  mockTimestampToDate,
-  mockTimestampToMillis,
-  mockTimestampFromDate,
-  mockTimestampFromMillis,
-  mockGet: query.mocks.mockGet,
-  mockOrderBy: query.mocks.mockOrderBy,
-  mockLimit: query.mocks.mockLimit,
-  mockOffset: query.mocks.mockOffset,
-  mockWhere: query.mocks.mockWhere,
+  ...query.mocks,
   ...transaction.mocks,
+  ...transaction.mocks,
+  ...fieldValue.mocks,
+  ...timestamp.mocks,
 };
